Rewrite ::DBI::Stories::process_extracted_story() to Python
Also reshuffle some stuff around to avoid cyclic dependency.
