+++ lib/MediaWords/DBI/Stories.pm
@@ -31,7 +31,8 @@ use List::Util;
 use MediaWords::DBI::Downloads;
 use MediaWords::DBI::Stories::ExtractorVersion;
 use MediaWords::DBI::Stories::ExtractorArguments;
-use MediaWords::DBI::Stories::Processed;
+use MediaWords::DBI::Stories::Postprocess;
+use MediaWords::DBI::Stories::Process;
 use MediaWords::Job::CLIFF::FetchAnnotation;
 use MediaWords::Job::NYTLabels::FetchAnnotation;
 use MediaWords::Languages::Language;
@@ -165,72 +166,9 @@ SQL
         MediaWords::DBI::Downloads::extract_and_create_download_text( $db, $download, $extractor_args );
     }
 
-    process_extracted_story( $db, $story, $extractor_args );
+    Mediawords::DBI::Stories::Process::process_extracted_story( $db, $story, $extractor_args );
 
     $db->commit if ( $use_transaction );
-}
-
-=head2 process_extracted_story( $db, $story, $extractor_args )
-
-Do post extraction story processing work by calling
-MediaWords::StoryVectors::update_story_sentences_and_language()
-
-=cut
-
-sub process_extracted_story($$$)
-{
-    my ( $db, $story, $extractor_args ) = @_;
-
-    my $stories_id = $story->{ stories_id } + 0;
-
-    MediaWords::StoryVectors::update_story_sentences_and_language( $db, $story, $extractor_args );
-
-    unless ( $extractor_args->no_tag_extractor_version() )
-    {
-        MediaWords::DBI::Stories::ExtractorVersion::update_extractor_version_tag( $db, $story );
-    }
-
-    my $cliff     = MediaWords::Util::Annotator::CLIFF->new();
-    my $nytlabels = MediaWords::Util::Annotator::NYTLabels->new();
-
-    # Extract -> CLIFF -> NYTLabels -> mark_as_processed() chain
-    if ( $cliff->annotator_is_enabled() and $cliff->story_is_annotatable( $db, $stories_id ) )
-    {
-        # If CLIFF annotator is enabled, ::CLIFF::UpdateStoryTags will check
-        # whether NYTLabels annotator is enabled, and if it is, will pass the
-        # story further to NYTLabels. NYTLabels, in turn, will mark the story
-        # as processed.
-        TRACE "Adding story $stories_id to CLIFF annotation queue...";
-        MediaWords::Job::CLIFF::FetchAnnotation->add_to_queue( { stories_id => $stories_id } );
-    }
-    else
-    {
-
-        TRACE "Won't add $stories_id to CLIFF annotation queue because it's not annotatable with CLIFF";
-
-        if ( $nytlabels->annotator_is_enabled() and $nytlabels->story_is_annotatable( $db, $stories_id ) )
-        {
-
-            # If CLIFF annotator is disabled, pass the story to NYTLabels
-            # annotator which, if run, will mark the story as processed
-            TRACE "Adding story $stories_id to NYTLabels annotation queue...";
-            MediaWords::Job::NYTLabels::FetchAnnotation->add_to_queue( { stories_id => $stories_id } );
-
-        }
-        else
-        {
-
-            TRACE "Won't add $stories_id to NYTLabels annotation queue because it's not annotatable with NYTLabels";
-
-            # If neither of the annotators are enabled, mark the story as processed ourselves
-            TRACE "Marking the story as processed...";
-            unless ( MediaWords::DBI::Stories::Processed::mark_as_processed( $db, $stories_id ) )
-            {
-                die "Unable to mark story ID $stories_id as processed";
-            }
-
-        }
-    }
 }
 
 =head2 _restore_download_content( $db, $download, $story_content )
