--- lib/MediaWords/DBI/Stories/Processed.pm
@@ -1,11 +1,0 @@
-package MediaWords::DBI::Stories::Processed;
-
-use strict;
-use warnings;
-
-use Modern::Perl "2015";
-use MediaWords::CommonLibs;
-
-import_python_module( __PACKAGE__, 'mediawords.dbi.stories.processed' );
-
-1;
